+++ composer.json
@@ -17,8 +17,7 @@
     "require": {
         "php": ">=5.4.0",
         "symfony/process": "2.*",
-        "aws/aws-sdk-php": "~2.4",
-        "league/flysystem": ">=0.3.1"
+        "league/flysystem": "~0.5"
     },
     "require-dev": {
         "phpunit/phpunit": "3.7.*",
@@ -29,7 +28,8 @@
         "illuminate/console": "4.x",
         "rackspace/php-opencloud": "1.9.*",
         "dropbox/dropbox-sdk": "1.1.*",
-        "phpspec/phpspec": "~2.0"
+        "phpspec/phpspec": "~2.0",
+        "aws/aws-sdk-php": "~2.4"
     },
     "config": {
         "bin-dir": "bin"
