+++ spec/BigName/BackupManager/Tasks/Compression/DecompressFileSpec.php
@@ -3,7 +3,7 @@
 namespace spec\BigName\BackupManager\Tasks\Compression;
 
 use BigName\BackupManager\Compressors\Compressor;
-use BigName\BackupManager\ShellProcessing\ShellProcessor;
+use BigName\BackupManager\Shell\ShellProcessor;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
