+++ src/BackupManagerException.php
@@ -1,0 +1,3 @@
+<?php namespace BigName\BackupManager;
+
+interface BackupManagerException {}
