+++ src/Compressors/CompressorTypeNotSupported.php
@@ -1,7 +1,9 @@
 <?php namespace BigName\BackupManager\Compressors;
 
+use BigName\BackupManager\BackupManagerException;
+
 /**
  * Class CompressorTypeNotSupported
  * @package BigName\BackupManager\Compressors
  */
-class CompressorTypeNotSupported extends \Exception {}
+class CompressorTypeNotSupported extends \Exception implements BackupManagerException {}
