+++ src/Config/ConfigFieldNotFound.php
@@ -1,7 +1,9 @@
 <?php namespace BigName\BackupManager\Config;
 
+use BigName\BackupManager\BackupManagerException;
+
 /**
  * Class ConfigFieldNotFound
  * @package BigName\BackupManager\Config
  */
-class ConfigFieldNotFound extends \Exception {}
+class ConfigFieldNotFound extends \Exception implements BackupManagerException {}
