+++ src/Config/ConfigFileNotFound.php
@@ -1,7 +1,9 @@
 <?php namespace BigName\BackupManager\Config;
 
+use BigName\BackupManager\BackupManagerException;
+
 /**
  * Class ConfigFileNotFound
  * @package BigName\BackupManager\Config
  */
-class ConfigFileNotFound extends \Exception {}
+class ConfigFileNotFound extends \Exception implements BackupManagerException {}
