+++ src/Config/ConfigNotFoundForConnection.php
@@ -1,7 +1,9 @@
 <?php namespace BigName\BackupManager\Config;
 
+use BigName\BackupManager\BackupManagerException;
+
 /**
  * Class ConfigNotFoundForConnection
  * @package BigName\BackupManager\Config
  */
-class ConfigNotFoundForConnection extends \Exception {}
+class ConfigNotFoundForConnection extends \Exception implements BackupManagerException {}
