+++ src/Databases/DatabaseTypeNotSupported.php
@@ -1,7 +1,9 @@
 <?php namespace BigName\BackupManager\Databases;
 
+use BigName\BackupManager\BackupManagerException;
+
 /**
  * Class DatabaseTypeNotSupported
  * @package BigName\BackupManager\Databases
  */
-class DatabaseTypeNotSupported extends \Exception {}
+class DatabaseTypeNotSupported extends \Exception implements BackupManagerException {}
