+++ src/Filesystems/FilesystemTypeNotSupported.php
@@ -1,7 +1,9 @@
 <?php namespace BigName\BackupManager\Filesystems;
 
+use BigName\BackupManager\BackupManagerException;
+
 /**
  * Class FilesystemTypeNotSupported
  * @package BigName\BackupManager\Filesystems
  */
-class FilesystemTypeNotSupported extends \Exception {}
+class FilesystemTypeNotSupported extends \Exception implements BackupManagerException {}
