+++ src/Shell/ShellProcessFailed.php
@@ -1,0 +1,9 @@
+<?php namespace BigName\BackupManager\Shell;
+
+use BigName\BackupManager\BackupManagerException;
+
+/**
+ * Class ShellProcessFailed
+ * @package BigName\BackupManager\Shell
+ */
+class ShellProcessFailed extends \Exception implements BackupManagerException {}
