+++ src/Shell/ShellProcessor.php
@@ -1,6 +1,7 @@
-<?php namespace BigName\BackupManager\ShellProcessing;
+<?php namespace BigName\BackupManager\Shell;
 
 use Symfony\Component\Process\Process;
+use Symfony\Component\Process\Exception\LogicException;
 
 /**
  * Class CommandProcessor
@@ -9,7 +10,7 @@ use Symfony\Component\Process\Process;
 class ShellProcessor
 {
     /**
-     * @var \Symfony\Component\Process\Process
+     * @var Process
      */
     private $process;
 
@@ -22,20 +23,17 @@ class ShellProcessor
     }
 
     /**
-     * @param $command
+     * @param  string $command
      * @throws ShellProcessFailed
-     * @throws \Symfony\Component\Process\Exception\LogicException
+     * @throws LogicException
      */
     public function process($command)
     {
-        if (empty($command)) {
+        if (empty($command))
             return;
-        }
 
-        $this->process->setCommandLine($command);
-        $this->process->run();
-        if ( ! $this->process->isSuccessful()) {
+        $this->process->setCommandLine($command)->run();
+        if ( ! $this->process->isSuccessful())
             throw new ShellProcessFailed($this->process->getErrorOutput());
-        }
     }
 }
