--- src/ShellProcessing/ShellProcessFailed.php
@@ -1,7 +1,0 @@
-<?php namespace BigName\BackupManager\ShellProcessing;
-
-/**
- * Class ShellProcessFailed
- * @package BigName\BackupManager\ShellProcessing
- */
-class ShellProcessFailed extends \Exception {}
