+++ src/Tasks/Compression/CompressFile.php
@@ -2,7 +2,8 @@
 
 use BigName\BackupManager\Tasks\Task;
 use BigName\BackupManager\Compressors\Compressor;
-use BigName\BackupManager\ShellProcessing\ShellProcessor;
+use BigName\BackupManager\Shell\ShellProcessor;
+use BigName\BackupManager\Shell\ShellProcessFailed;
 
 /**
  * Class CompressFile
@@ -15,11 +16,11 @@ class CompressFile implements Task
      */
     private $sourcePath;
     /**
-     * @var \BigName\BackupManager\ShellProcessing\ShellProcessor
+     * @var ShellProcessor
      */
     private $shellProcessor;
     /**
-     * @var \BigName\BackupManager\Compressors\Compressor
+     * @var Compressor
      */
     private $compressor;
 
@@ -36,7 +37,7 @@ class CompressFile implements Task
     }
 
     /**
-     * @throws \BigName\BackupManager\ShellProcessing\ShellProcessFailed
+     * @throws ShellProcessFailed
      */
     public function execute()
     {
